+++ hubdle-project-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/documentation/changelog/AddChangelogItemTask.kt
@@ -102,7 +102,7 @@ constructor(
     }
 
     public companion object {
-        public const val name: String = "addChangelogItem"
+        public const val NAME: String = "addChangelogItem"
     }
 }