+++ hubdle-project-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/documentation/changelog/ApplyFormatChangelogTask.kt
@@ -22,6 +22,6 @@ public abstract class ApplyFormatChangelogTask : DefaultTask() {
     }
 
     public companion object {
-        public const val name: String = "applyFormatChangelog"
+        public const val NAME: String = "applyFormatChangelog"
     }
 }