+++ hubdle-project-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/documentation/changelog/GenerateChangelogHtmlTask.kt
@@ -35,7 +35,7 @@ constructor(
     }
 
     public companion object {
-        public const val name: String = "generateChangelogHtml"
+        public const val NAME: String = "generateChangelogHtml"
     }
 }